Deploy should fail if rake db:migrate fails
Previously, if `rake db:migrate` failed during the deploy script, the
deploy would continue as if nothing has gone wrong. It seems to me that
the deploy should fail in this scenario.

Fix by adding the `-x` flag to `heroku run`, which causes the heroku
command to passthrough the exit code of the remote command.