Rotate log files to keep them from growing forever
The development.log and test.log files accumulate logs whenever the app
and its tests are run, database migrations are performed, etc. Over the
course of a few months of a developer working on a Rails project, these
files can easily grow to exceed 1GB.

Rails has a built-in `log:clear` rake task to delete the log files, but
developers have to remember to run this periodically. The task is also
executed by `bin/setup` and `bin/update`, but again these are unlikely
to be used once a developer is set up.

This commit automatically detects the size of the log files upon startup
and rotates them (with one backup) to keep them under 5MB each.