Move generator config into application.rb
The official Rails documentation (as well docs for factory_bot_rails and
others gems) recommend placing generator config in
`config/application.rb`.

This is actually important because of load order. Initializers are
executed later in the boot process, after the generators provided by
gems are loaded. Thus code in an initializer could be executed too late/

Case in point: the `g.factory_bot` setting, if set in an initializer,
comes too late to control the loading of factory_bot_rails. In other
words, `g.factory_bot false` would not actually turn off the FactorBot
generator as expected.

When the same code is placed in `application.rb` instead, it has the
intended effect.

This commit also renames `g.factory_girl` to `g.factory_bot`.