Use `heroku local` instead of foreman
The README recommended `foreman start` as the way to run the app, but in
practice this does not work: the `release` process in the `Procfile`
exits, which causes foreman to shutdown all other processes:

```
13:44:10 release.1 | exited with code 0
13:44:10 system    | sending SIGTERM to all processes
13:44:10 web.1     | - Gracefully stopping, waiting for requests to finish
13:44:10 web.1     | === puma shutdown: 2017-11-27 13:44:10 -0800 ===
13:44:10 web.1     | - Goodbye!
13:44:10 web.1     | terminated by SIGTERM
```

A workaround is to ignore the release process, like this:

```
foreman start -m all=1,release=0
```

But there is an even easier solution, which is to use `heroku local`.
This is essentially a wrapper of `node-foreman` that automatically
ignores the `release` process.

This commit removes the `foreman` gem and updates the README to explain
that `heroku local` is the suggested way to run the app locally.