Only run bundler-audit on master/development
Dealing with security updates in the middle of feature development can
be distracting, especially when the same gem has to be updated in every
branch. It's better to update the gem once, merge that into master, and
then all the other features will benefit once they are later merged into
master as well.

With that workflow in mind, remove the `bundler-audit` job from CircleCI
for feature branches in order to minimize interruptions. But still run
in the master and development branches because security is important and
we don't want to go to production without addressing audit issues.