Allow headless Chrome to be disabled (opt-in)
When writing or troubleshooting specs, it is useful to run the regular
(i.e. "headed") version of Chrome so you can see what is being rendered
and use the Chrome developer tools. However we still want headless
Chrome to be the default, because headless specs run faster.

Allow opting into headed mode via an ENV var.