Specify bundle config via ENV; various cleanup
This commit changes the bundler-related steps that we use in Circle CI
to bring them up to speed with practices we have in other C5 projects.

First, add a `bundle check` step to speed things up. For large Gemfiles
with many sources, `bundle install` can lag a few seconds even if there
is nothing to install. Whereas `bundle check` is much faster and can
eliminate the need to run `bundle install` in many cases.

Note that both `check` and `install` require a `--path` option. Rather
than repeat this, specify it once via the `BUNDLE_PATH` environment
variable. For consistency, specify `BUNDLE_JOBS` there as well.

Also, the `--force` option is only needed when gems are installed
globally (i.e. no path option is specified). Since we are using a path,
`--force` does nothing and can be removed.

Finally, the Circle CI documentation recommends that `{{ arch }}` always
be included in the cache key, even for fallbacks. Add it.