Explicitly list the Rails frameworks to load
Almost no app generated from raygun-rails will use the entire suite of
frameworks that are bundled with Rails. It was not obvious how to opt
out of the ones that are not needed.

This commit replaces `"rails/all"` with the explicit list of frameworks
that make up Rails 6.0, so that a developer can easily delete or comment
out the ones they do not wish to use. In our case we are using RSpec
instead of TestUnit, so we can comment out the TestUnit framework.