Remove bootstrap
Problem
-------

Bootstrap, though it can help to get things rolling, it is not ideal in
the modern day where we often move rapidly to a Javascript
library/framework that has CSS-in-JS or some other scheme to manage
styling.

Including bootstrap includes (implicitly) it's own reset CSS that is
fine at first.  But if you later decide to remove Bootstrap, you may
find that your components no longer look right because you've lost that
implicit reset.

Solution
--------

Remove bootstrap

Changes
-------

* Remove bootstrap from package.json
* update the stylesheets we have to be a bit simpler but still rendering
a header and footer