Require Node 10+
Node 8 is now EOL and many dependencies of this project no longer
support it. So effectively this means we require Node 10 or higher.
Putting an "engine" in the `package.json` ensures that developers will
(and e.g. CI) will fail loudly with a useful error message if they try
to build the project using an outdated version of Node by mistake.