Regenerate config for compatibility w annotate 3.x
The 3.x version of annotate added some new configuration options that
are needed for it to work properly. Most notably if `"models"=>"true"`
is missing it will not annotate anything!

Rerun the annotate installer to get the latest recommend config. The
only differences for our config versus the defaults are:

- `"exclude_tests" => "true"` (don't annotate spec files)
- `"sort" => "true"` and `"classified_sort" => "false"` (sort
  columns alphabetically and not by column type)