Add rubocop-rspec and inherit new c5-conventions (#436)
Problem:

We'd like to add rubocop-rspec to our projects going forward. The
c5-conventions has a recommended config for rubocop-rspec that we can
opt into.

Solution:

Inherit the rubocop-rspec conventions, along with all the other latest
conventions, from the new `main` branch of the c5-conventions repo.

Add `rubocop-rpsec` to the Gemfile.