Remove Heroku CLI requirement for running locally (#438)
* Remove Heroku CLI requirement for running locally

Using the Heroku CLI to start the app locally is a pretty heavy
requirement. It would be great if raygun projects were a little more
portable so that you can run the app locally without using Heroku.

These days, the Heroku CLI is written in Typescript and delegates to the
`foreman` node package behind the scenes for its `heroku local` command.
So we can slim down the requirements for running this project quite
significantly and make it more platform agnostic by using foreman
directly.

I left the Heroku CLI requirement listed in the README because other
parts of the project do still need Heroku, like the database export
scripts in `bin/`. These would be easy to remove for projects decide to
switch from Heroku to a different platform for deployment. The start
script from running locally is a bit trickier to replace, which is why
I think it is worth making it more generic. Hence this PR.

* Clarify Heroku CLI is used for deployment only