Use `concurrently` for better DX than foreman (#535)
Up until now, we have used foreman (`nf`) or `heroku local` pointing at
a `Procfile.dev` to run the app when someone typed `yarn start`. This
leads to some problems, though. The output would be strangely buffered
so that sometimes output you'd expect to see from Rails would be missing
or would lag in the console output, but would show up in
`log/development.log`.

Furthermore it could be very difficult to use `byebug` for interactive
debugging in the console due to these buffering issues.

As a solution, drop foreman and use the `concurrently` utility instead,
which gives us much more control.

- `--raw` disables any sort of buffering
- `--kill-others-on-fail` means that if one process fails (e.g. the
  rails server can't boot because the port is taken), the whole
  `yarn start` command will exit
- stdin is sent to the first process by default (rails), so `byebug`
  works

`concurrently` is a popular and well-supported project:
https://snyk.io/advisor/npm-package/concurrently