Fix broken .nvmrc symlink (#604)
The `.nvmrc` file is supposed to be a symlink that points to
`.node-version`. That way we can update the node version in one place,
and various node version managers will be able to access it (nvm uses
`.nvmrc`, all others support `.node-version`).

For some reason the `.nvmrc` was a plain file instead of a symlink. This
caused `nvm use` to fail.

Fix by create the symlink appropriately.