Eager load in CI; remove non-functional ActionCable files (#632)
* Eager load the app when testing under CI

This is a port of a new recommendation in Rails 7, but we can safely
include it in our Rails 6.1 template.

Basically, there can be load-ordering issues that only happen when eager
loading is turned on. Currently, we disable eager loading in dev and
test. That means we can get an unpleasant surprise in production.

Address this by turning eager loading on in CI. That guarantees we are
testing the eager load scenario prior to production, but still keeping
it off by default (for speed) in development and local testing.

https://github.com/rails/rails/pull/43508/files

* Remove non-functional ActionCable files

We do not require `action_cable/engine` in `application.rb`, so that
means these ActionCable files will prevent the application from booting
in production.

If we are not going to enable ActionCable and still want an app that
works out of the box, we need to remove these files.