Fix babel "loose" config warnings
Problem

Running webpack prints these warnings:

```
Though the "loose" option was set to "false" in your @babel/preset-env config, it will not be used for @babel/plugin-proposal-private-property-in-object since the "loose" mode option was set to "true" for @babel/plugin-proposal-private-methods.
The "loose" option must be the same for @babel/plugin-proposal-class-properties, @babel/plugin-proposal-private-methods and @babel/plugin-proposal-private-property-in-object (when they are enabled): you can silence this warning by explicitly adding
  ["@babel/plugin-proposal-private-property-in-object", { "loose": true }]
to the "plugins" section of your Babel config.
```

Solution

This is due to the `babel.config.js` that is generated by webpacker. The
problem has been fixed upstream.

Solution is to regenerate `babel.config.js` using the latest version of
webpacker that we are using (5.4.3).

Steps to verify

1. Delete `node_modules`.
2. Run `bin/setup`.
3. Run `bin/webpack`.
4. Confirm that no "loose" warnings are printed.