Fix "no such middleware" error in production mode (#712)
Problem:

If the app is run in production and `RAILS_SERVE_STATIC_FILES` is
**not** set, the app fails to boot with this error:

```
No such middleware to insert after: ActionDispatch::Static
config/environment.rb:5:in `<main>'
Tasks: TOP => assets:precompile => webpacker:compile => environment
(See full trace by running task with --trace)
```

That's because the `ActionDispatch::Static` middleware only exists in
production if the `config.public_file_server` value is set to true. If
`RAILS_SERVE_STATIC_FILES` is not present, it is set to false, and thus
the middleware doesn't exist.

Solution:

Move the `Rack::Deflater` setup into `production.rb` alongside the
`config.public_file_server` setting, so that the middlware is only
inserted when the latter is set to true.